List lag of prior-list intrusions

For intrusions of items from previous lists, the lag indicating how many lists back the item was presented.

Usage

pli_list_lag(data, max_lag)

Arguments

data

Merged study and recall data. Lists must be numbered starting from 1 and all lists must be included.

max_lag

Maximum list lag to consider. The intial max_lag lists for each subject will be excluded so that all considered lags are possible for all included lists.

Value

Results with subject, list_lag, count, per_list, and prob columns. Count indicates the number of intrusions with that list lag, while per_list indicates the number of intrusions per list. The prob column indicates the probability within each subject of a given included prior-list intrusion occurring at that lag.

Examples

raw <- sample_data("Morton2013")
data <- merge_free_recall(raw)
stats <- pli_list_lag(data, max_lag = 3)
head(stats)
#>   subject list_lag count   per_list       prob
#> 1       1        1     7 0.15555556 0.25925926
#> 2       1        2     5 0.11111111 0.18518519
#> 3       1        3     0 0.00000000 0.00000000
#> 4       2        1     9 0.20000000 0.19148936
#> 5       2        2     2 0.04444444 0.04255319
#> 6       2        3     1 0.02222222 0.02127660